OnStart up Tasks 6

+++ src/main/java/com/stackroute/muzixapp/service/MuzixServiceImpl.java
@@ -1,15 +1,22 @@
 package com.stackroute.muzixapp.service;
 
 import com.stackroute.muzixapp.domain.Track;
+import com.stackroute.muzixapp.exceptions.TrackAlreadyExistsException;
+import com.stackroute.muzixapp.exceptions.TrackNotFoundException;
 import com.stackroute.muzixapp.repository.MuzixRepository;
+import org.apache.commons.logging.Log;
+import org.apache.commons.logging.LogFactory;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.boot.CommandLineRunner;
+import org.springframework.context.ApplicationListener;
+import org.springframework.context.event.ContextRefreshedEvent;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
 import java.util.Optional;
 
 @Service
-public class MuzixServiceImpl implements MuzixService {
+public class MuzixServiceImpl implements MuzixService , ApplicationListener<ContextRefreshedEvent> ,CommandLineRunner{
     MuzixRepository muzixRepository;
 
 
@@ -17,16 +24,26 @@ public class MuzixServiceImpl implements MuzixService {
     public MuzixServiceImpl(MuzixRepository muzixRepository){
         this.muzixRepository=muzixRepository;
     }
+
+
     @Override
-    public Track saveTrack(Track track)  {
-
+    public Track saveTrack(Track track)  throws TrackAlreadyExistsException {
+        if(muzixRepository.existsById(track.getId())){
+            throw new TrackAlreadyExistsException("User already exists!");
+        }
            Track savedTrack = muzixRepository.save(track);
-
+        if(savedTrack == null) {
+            throw new TrackAlreadyExistsException("User already exists!");
+        }
         return savedTrack;
     }
 
     @Override
-    public void deleteById(int id) {
+    public void deleteById(int id) throws TrackNotFoundException {
+        Optional<Track> userId = muzixRepository.findById(id);
+       if(!userId.isPresent()){
+            throw new TrackNotFoundException("Track not found!");
+        }
         muzixRepository.deleteById(id);
     }
 
@@ -52,6 +69,21 @@ public class MuzixServiceImpl implements MuzixService {
     @Override
     public List<Track> getAllTracks() {
         return muzixRepository.findAll();
+    }
+    // For Application Listner
+    @Override
+    public void onApplicationEvent(ContextRefreshedEvent contextRefreshedEvent) {
+        muzixRepository.save(new Track(1, "all of me", "Sudeep"));
+        muzixRepository.save(new Track(2, "DNA",  "Ravi"));
+    }
+
+
+    @Override
+    public void run(String... args) throws Exception {
+        final Log logger = LogFactory.getLog(getClass());
+        logger.info("Application Started ..............!");
+
+
     }
 
 }
